Move generic interfaces to a new file Remove ununsed library import Remove logging Linting Remove "pewpew" from objects
MIMIR-766
