+++ src/main/resources/services/websocket/websocket.es6
@@ -1,10 +1,7 @@
-const {
-  run
-} = __non_webpack_require__('/lib/xp/context')
 const ws = __non_webpack_require__('/lib/wsUtil')
 const convert = __non_webpack_require__('/lib/convert/convert')
 const dashboard = __non_webpack_require__('/lib/ssb/dataset/dashboard')
-const auth = __non_webpack_require__('/lib/xp/auth')
+
 ws.openWebsockets(exports) // make the server ready for socket connections
 
 // use SocketEmitter expansion to handle this
